+++ apps/docs/src/components/demos/features/CountryBlacklist.tsx
@@ -1,33 +1,37 @@
-import React, { useState } from 'react';
+import React, { useMemo, useState } from 'react';
 import { CountryDropdown, RegionDropdown } from 'react-country-region-selector';
 
 const CountryBlacklist = () => {
   const [country, setCountry] = useState('');
   const [region, setRegion] = useState('');
+  const blacklist = useMemo(
+    () => [
+      'AF',
+      'AX',
+      'AL',
+      'DZ',
+      'AS',
+      'AD',
+      'AO',
+      'AI',
+      'AQ',
+      'AG',
+      'AR',
+      'AM',
+      'AW',
+      'AU',
+      'AT',
+      'AZ',
+    ],
+    []
+  );
 
   return (
     <>
       <CountryDropdown
         value={country}
         onChange={(val) => setCountry(val)}
-        blacklist={[
-          'AF',
-          'AX',
-          'AL',
-          'DZ',
-          'AS',
-          'AD',
-          'AO',
-          'AI',
-          'AQ',
-          'AG',
-          'AR',
-          'AM',
-          'AW',
-          'AU',
-          'AT',
-          'AZ',
-        ]}
+        blacklist={blacklist}
       />
       <RegionDropdown
         country={country}