+++ apps/docs/src/components/demos/features/CountryWhitelist.tsx
@@ -1,16 +1,17 @@
-import React, { useState } from 'react';
+import React, { useMemo, useState } from 'react';
 import { CountryDropdown, RegionDropdown } from 'react-country-region-selector';
 
 const CountryWhitelist = () => {
   const [country, setCountry] = useState('');
   const [region, setRegion] = useState('');
+  const whitelist = useMemo(() => ['GB', 'US', 'CA'], []);
 
   return (
     <>
       <CountryDropdown
         value={country}
         onChange={(val) => setCountry(val)}
-        whitelist={['GB', 'US', 'CA']}
+        whitelist={whitelist}
       />
       <RegionDropdown
         country={country}